requirements.txt
streamlit
pandas
openpyxl
streamlit-ace
google-cloud-bigquery
streamlit-tree-select
google-auth
google-cloud-aiplatform
google-genai
dependency-injector
PyYAML
PyQt6
